+++ crates/bevy_render/src/mesh/mod.rs
@@ -2,6 +2,7 @@ pub mod allocator;
 pub mod material_bind_group;
 pub mod render;
 pub mod skin;
+pub mod pipeline;
 pub mod util;
 use crate::{
     render_asset::{PrepareAssetError, RenderAsset, RenderAssetPlugin, RenderAssets},
